modules/07-ransomware-sim/requirements.txt
PyYAML
pycryptodome
